requirements.txt
ai-django-core==1.2.7
asn1crypto==1.0.1
atomicwrites==1.3.0
attrs==19.3.0
bleach==3.1.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
cryptography==2.7
dataclasses==0.6
Django==2.2.6
django-cors-headers==3.1.1
django-crispy-forms==1.8.0
django-filter==2.2.0
django-nexmo==1.0.1
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
djangorestframework==3.10.3
djongo==1.2.36
freezegun==0.3.12
future==0.18.1
idna==2.8
importlib-metadata==0.23
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mock==3.0.5
more-itertools==7.2.0
openapi-codec==1.3.2
packaging==19.2
Pillow==6.2.0
pluggy==0.13.0
py==1.8.0
pycparser==2.19
pymongo==3.9.0
pyparsing==2.4.2
pytest==5.2.1
pytest-cov==2.8.1
pytest-django==3.6.0
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
zipp==0.6.0
